Move functions and orbital functions to sub-module
This will enable future functions to be grouped logically, and may also
consider breaking up existing functions in future.
